src/main/java/com/biblioteca/controller/CheckoutController.java
package com.biblioteca.controller;

import com.biblioteca.service.CheckoutService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class CheckoutController {

    private CheckoutService checkoutService;

    @Autowired
    public CheckoutController(CheckoutService checkoutService) {
        this.checkoutService = checkoutService;
    }

    @RequestMapping(value = "/books/{id}/checkout")
    public void checkoutBook(@PathVariable("id") String id) {
        checkoutService.checkoutBook(id);
    }
}
